+++ prisma/seed.ts
@@ -1,3 +1,4 @@
+import { faker } from "@faker-js/faker"
 import { PrismaPg } from "@prisma/adapter-pg"
 import { PrismaClient } from "../src/generated/prisma/client"
 
@@ -10,16 +11,50 @@ const prisma = new PrismaClient({
 	adapter,
 })
 
-const userData = [
-	{
-		email: "ibrahimolayinkaa@gmail.com",
-		phone: "+2347065643303",
-		firstName: "Olayinka",
-		lastName: "Ibrahim",
-	},
-]
+function generateNigerianPhone(): string {
+	const prefixes = [
+		"803",
+		"806",
+		"813",
+		"816",
+		"703",
+		"706",
+		"903",
+		"906",
+		"805",
+		"807",
+		"811",
+		"815",
+		"905",
+		"802",
+		"809",
+		"812",
+		"902",
+		"909",
+		"817",
+		"818",
+		"908",
+		"701",
+		"708",
+	]
+	const randomPrefix = faker.helpers.arrayElement(prefixes)
+	const remainingDigits = faker.string.numeric(7) // Generates 7 random digits
+	return `+234${randomPrefix}${remainingDigits}`
+}
 
-export async function main() {
+const userData = Array.from({ length: 10 }).map(() => {
+	const firstName = faker.person.firstName()
+	const lastName = faker.person.lastName()
+
+	return {
+		firstName,
+		lastName,
+		email: faker.internet.email({ firstName, lastName }).toLowerCase(),
+		phone: generateNigerianPhone(),
+	}
+})
+
+async function main() {
 	for (const u of userData) {
 		await prisma.user.create({
 			data: u,