+++ src/modules/user/user.repository.ts
@@ -1,6 +1,20 @@
 import { injectable } from "inversify"
-// import type { IUserRepository } from "./user.contracts";
+import { prisma } from "@/core/database/db"
+import type { IUserRepository } from "./user.types"
 
 @injectable()
-export class UserRepository {}
-// export class UserRepository implements IUserRepository {}
+export class UserRepository implements IUserRepository {
+	async findByPhoneWithKyc(phone: string) {
+		return prisma.user.findUnique({
+			where: { phone },
+			select: {
+				id: true,
+				kyc: {
+					select: {
+						completedProfile: true,
+					},
+				},
+			},
+		})
+	}
+}